pythoncrm/requirements.txt
apturl==0.5.2
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
certifi==2018.1.18
chardet==3.0.4
checkbox-support==0.22
command-not-found==0.3
configparser==3.5.0
crypto==1.4.1
cryptography==1.2.3
cytoolz==0.9.0
defer==1.0.6
Django==1.8
django-filter==1.1.0
djangorestframework==3.5.3
djangorestframework-jwt==1.11.0
eth-abi==0.5.0
eth-keys==0.1.0b4
eth-tester==0.1.0b11
eth-utils==0.7.4
feedparser==5.1.3
gevent==1.2.2
greenlet==0.4.13
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.6
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lru-dict==1.1.6
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.11
MarkupSafe==0.23
Naked==0.1.31
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
ptyprocess==0.5
py-solc==2.1.0
pyasn1==0.1.9
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.6.1
PyMySQL==0.8.0
pyparsing==2.0.3
pysha3==1.0.2
python-apt==1.1.0b1
python-debian==0.1.27
python-systemd==231
pytz==2017.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.3.0
requests==2.18.4
rlp==0.6.0
semantic-version==2.6.0
sessioninstaller==0.0.0
shellescape==3.4.1
six==1.10.0
system-service==0.3
toolz==0.9.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.0
web3==3.16.5
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
xlwt==1.3.0
